# src/Vibetech.Educat.Web/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Vibetech Educat API",
    "description": "API для платформы репетиторов и студентов",
    "contact": {
      "name": "Vibetech",
      "email": "support@vibetech.ru"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Авторизация пользователя",
        "description": "Авторизует пользователя в системе и возвращает JWT-токен",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Пользователь успешно авторизован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Неверные учетные данные"
          }
        }
      }
    },
    "/api/Auth/register/student": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Регистрация студента",
        "description": "Регистрирует нового студента в системе",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentRegisterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StudentRegisterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StudentRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Студент успешно зарегистрирован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка регистрации"
          }
        }
      }
    },
    "/api/Auth/register/teacher": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Регистрация преподавателя",
        "description": "Регистрирует нового преподавателя в системе. Поля SubjectIds и PreparationProgramIds не обязательны и могут быть добавлены позже.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeacherRegisterRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TeacherRegisterRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TeacherRegisterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Преподаватель успешно зарегистрирован",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Ошибка регистрации"
          }
        }
      }
    },
    "/api/Dictionary/subjects": {
      "get": {
        "tags": [
          "Dictionary"
        ],
        "summary": "Получение списка предметов",
        "description": "Возвращает список всех доступных предметов",
        "responses": {
          "200": {
            "description": "Список предметов",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubjectDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubjectDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubjectDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Dictionary"
        ],
        "summary": "Добавление предмета",
        "description": "Добавляет новый предмет в систему",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSubjectRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSubjectRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddSubjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Предмет успешно добавлен",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubjectDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Dictionary/preparation-programs": {
      "get": {
        "tags": [
          "Dictionary"
        ],
        "summary": "Получение списка программ подготовки",
        "description": "Возвращает список всех доступных программ подготовки",
        "responses": {
          "200": {
            "description": "Список программ подготовки",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreparationProgramDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreparationProgramDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PreparationProgramDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Dictionary"
        ],
        "summary": "Добавление программы подготовки",
        "description": "Добавляет новую программу подготовки в систему",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPreparationProgramRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AddPreparationProgramRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AddPreparationProgramRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Программа подготовки успешно добавлена",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PreparationProgramDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreparationProgramDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreparationProgramDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AddPreparationProgramRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AddSubjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "token": {
            "type": "string",
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        },
        "additionalProperties": false
      },
...
      "LoginRequest": {
        "required": [
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "login": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "PreparationProgramDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "StudentRegisterRequest": {
        "required": [
          "birthDate",
          "confirmPassword",
          "contactInfo",
          "firstName",
          "gender",
          "lastName",
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "login": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "maxLength": 100,
            "minLength": 6,
            "type": "string"
          },
          "confirmPassword": {
            "minLength": 1,
            "type": "string"
          },
          "lastName": {
            "minLength": 1,
            "type": "string"
          },
          "firstName": {
            "minLength": 1,
            "type": "string"
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "description": "Дата рождения в формате YYYY-MM-DD",
            "format": "date",
            "example": "2025-05-28"
          },
          "gender": {
            "minLength": 1,
            "type": "string"
          },
          "contactInfo": {
            "minLength": 1,
            "type": "string"
          },
          "photoBase64": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SubjectDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TeacherRegisterRequest": {
        "required": [
          "birthDate",
          "confirmPassword",
          "contactInfo",
          "education",
          "experienceYears",
          "firstName",
          "gender",
          "hourlyRate",
          "lastName",
          "login",
          "password"
        ],
        "type": "object",
        "properties": {
          "login": {
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "maxLength": 100,
            "minLength": 6,
            "type": "string"
          },
          "confirmPassword": {
            "minLength": 1,
            "type": "string"
          },
          "lastName": {
            "minLength": 1,
            "type": "string"
          },
          "firstName": {
            "minLength": 1,
            "type": "string"
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "description": "Дата рождения в формате YYYY-MM-DD",
            "format": "date",
            "example": "2025-05-28"
          },
          "gender": {
            "minLength": 1,
            "type": "string"
          },
          "contactInfo": {
            "minLength": 1,
            "type": "string"
          },
          "photoBase64": {
            "type": "string",
            "nullable": true
          },
          "education": {
            "minLength": 1,
            "type": "string"
          },
          "preparationProgramIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          },
          "experienceYears": {
            "type": "integer",
            "format": "int32"
          },
          "hourlyRate": {
            "maximum": 10000,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "subjectIds": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "lastName": {
            "type": "string",
            "nullable": true
          },
          "firstName": {
            "type": "string",
            "nullable": true
          },
          "middleName": {
            "type": "string",
            "nullable": true
          },
          "birthDate": {
            "type": "string",
            "description": "Дата рождения в формате YYYY-MM-DD",
            "format": "date",
            "example": "2025-05-28"
          },
          "gender": {
            "type": "string",
            "nullable": true
          },
          "contactInfo": {
            "type": "string",
            "nullable": true
          },
          "isTeacher": {
            "type": "boolean"
          },
          "roles": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "photoBase64": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ],
...
}